+++ NEWS.md
@@ -6,6 +6,7 @@
 * 128 palettes have been added from the lisa package. (#35)
 * 16 palettes have been added from the IslamicArt package. (#36)
 * 9 palettes have been added from the basetheme package. (#41)
+* 8 palettes have been added from the vapeplot package. ($42)
 
 # paletteer 0.2.1